fix: bugs after merge

+++ src/serializer/nfs/fs_stat.rs
@@ -1,5 +1,5 @@
 use crate::serializer::nfs::files::file_attr;
-use crate::serializer::{option, u32};
+use crate::serializer::{option, u32, u64};
 use crate::vfs::fs_stat;
 use std::io;
 use std::io::Write;
@@ -7,12 +7,12 @@ use std::io::Write;
 #[allow(dead_code)]
 pub fn fs_stat_res_ok<S: Write>(dest: &mut S, arg: fs_stat::Success) -> io::Result<()> {
     option(dest, arg.root_attr, |attr, dest| file_attr(dest, attr))?;
-    u32(dest, arg.total_bytes)?;
-    u32(dest, arg.free_bytes)?;
-    u32(dest, arg.available_bytes)?;
-    u32(dest, arg.total_files)?;
-    u32(dest, arg.free_files)?;
-    u32(dest, arg.available_files)?;
+    u64(dest, arg.total_bytes)?;
+    u64(dest, arg.free_bytes)?;
+    u64(dest, arg.available_bytes)?;
+    u64(dest, arg.total_files)?;
+    u64(dest, arg.free_files)?;
+    u64(dest, arg.available_files)?;
     u32(dest, arg.invarsec)
 }
